refactor: split plain_vector_index.rs into a module
Split `plain_vector_index.rs` into a `plain_vector_index/` module:

- `mod.rs`: the `PlainVectorIndex` struct and its `new` constructor
- `read.rs`: `is_small_enough_for_unindexed_search` and the
  `VectorIndexRead` impl (search, telemetry, etc.)
- `lifecycle.rs`: the `VectorIndex` impl (`files`, `update_vector`)

Code was moved verbatim; only the per-file `use` lists were adjusted.
No behavior change.
